+ pr #35 tmux, gamedig, cleanup

+++ scripts/cleanImage.sh
@@ -3,12 +3,15 @@
 set -o errexit
 set -o nounset
 echo "[cleanImage] cleaning image"
-
+if [ -z "$LGSM_PATH" ]; then
+    exit 10
+fi
 
 apt-get autoremove -y
 apt-get clean -y
 rm -rf /var/lib/apt/lists/* >> /dev/null 2>&1 || true
-rm -rf "$LGSM_PATH/"*
-rm "/home/installMinimalDependencies.sh" \
-    "/home/installLGSM.sh" \
-    "/home/setupUser.sh" >> /dev/null 2>&1 || true
+rm -rf "$LGSM_PATH"/* /tmp/* /var/tmp/* || true
+rm "$LGSM_SCRIPTS/installMinimalDependencies.sh" \
+    "$LGSM_SCRIPTS/installLGSM.sh" \
+    "$LGSM_SCRIPTS/installGamedig.sh" \
+    "$LGSM_SCRIPTS/setupUser.sh" >> /dev/null 2>&1 || true
